move auth router group under landing router group
