+++ web-app/src/app/(landing)/(auth)/components/close-button.tsx
