+++ web-app/src/app/(landing)/(auth)/components/heroes.tsx
