+++ web-app/src/app/(landing)/layout.tsx
@@ -1,9 +1,8 @@
 import { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
 
-import MainFooter from "@/components/main-footer";
-
 import BreadcrumbNav from "./components/breadcrumb-nav";
+import Footer from "./components/footer";
 import Header from "./components/header";
 
 interface LandingLayoutProps {
@@ -28,7 +27,7 @@ export default function LandingLayout({ children }: LandingLayoutProps) {
       <Header />
       <BreadcrumbNav />
       <main className="flex-1">{children}</main>
-      <MainFooter />
+      <Footer />
     </div>
   );
 }
